Middleware tests

+++ tests/setup.ts
@@ -1,15 +1,18 @@
 import mongoose from 'mongoose';
 
 beforeAll(async () => {
-  await mongoose.connect(process.env.MONGO_URI_TEST as string);
+  if (process.env.MONGO_URI_TEST) {
+    await mongoose.connect(process.env.MONGO_URI_TEST);
+  }
 });
 
 afterAll(async () => {
-  if (mongoose.connection.db) {
-    await mongoose.connection.db.dropDatabase();
-  } else {
-    console.error('mongoose.connection.db is undefined');
+  if (process.env.MONGO_URI_TEST) {
+    if (mongoose.connection.db) {
+      await mongoose.connection.db.dropDatabase();
+    } else {
+      console.error('mongoose.connection.db is undefined');
+    }
   }
-  
   await mongoose.connection.close();
 });
